+++ core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.2.7]
 
 - Add polynomial multiproof features on core
+- Add kbucket pending timeout to kademlia configuration
 
 ## [1.2.6](https://github.com/availproject/avail-light/tree/avail-light-core-v1.2.6) - 2025-04-09